+++ public/index.html
@@ -78,12 +78,18 @@
         <!-- (6)登録ボタン -->
         <div class="form-group">
             <div class="col-md-offset-2 col-md-10">
-                    <button type="button" class="btn btn-default" id="post-btn">新規登録</button>
+                <button type="button" class="btn btn-default" id="post-btn">新規登録</button>
             </div>
         </div>
 
+        <!-- (6)登録ボタン -->
+        <div class="form-group">
+            <div class="col-md-offset-2 col-md-10">
+                <button type="button" class="btn btn-default" id="post-btn">ログイン</button>
+            </div>
+        </div>
     </form>
-    <script src="JS/main.js"></script>
+    <script src="../dist/js/bundle.js"></script>
     <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js"
         integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous">
     </script>
@@ -93,7 +99,6 @@
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"
         integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous">
     </script>
-
 </body>
 </body>
