+++ public/timeline.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8" />
     <title>横浜のツイッター</title>
-    <link rel='stylesheet' href='style.css' type='text/css' media='all' />
+    <link rel='stylesheet' href='./css/style.css' type='text/css' media='all' />
 </head>
 
 <body>
@@ -21,7 +21,7 @@
             <!-- 記事エリア -->
             <div class="twitter__block">
                 <figure>
-                    <img src="ryusei.jpg" />
+                    <img src="./img/ryusei.jpg" />
                 </figure>
                 <div class="twitter__block-text">
                     <div class="name">横浜流星<span class="name_reply">@ryuusei_tan</span></div>
@@ -29,7 +29,7 @@
                     <div class="text">
                         弟まじ神！かっこいい！
                         <div class="in-pict">
-                            <img src="kaito.jpg">
+                            <img src="img/kaito.jpg">
                         </div>
                     </div>
                     <div class="twitter__icon">
@@ -42,7 +42,7 @@
             <!-- 記事エリア -->
             <div class="twitter__block">
                 <figure>
-                    <img src="ryusei.jpg" />
+                    <img src="./img/ryusei.jpg" />
                 </figure>
                 <div class="twitter__block-text">
                     <div class="name">横浜流星<span class="name_reply">@ryuusei_tan</span></div>
@@ -50,7 +50,7 @@
                     <div class="text">
                         いやぁ、今日も俺イケメンだわ
                         <div class="in-pict">
-                            <img src="sample.jpg">
+                            <img src="./img/sample.jpg">
                         </div>
                     </div>
                     <div class="twitter__icon">
@@ -64,7 +64,7 @@
             <!-- 記事エリア -->
             <div class="twitter__block">
                 <figure>
-                    <img src="icon.png" />
+                    <img src="./img/icon.png" />
                 </figure>
                 <div class="twitter__block-text">
                     <div class="name">フェニックス<span class="name_reply">@kaikai3</span></div>
@@ -83,7 +83,7 @@
             <!-- 記事エリア -->
             <div class="twitter__block">
                 <figure>
-                    <img src="ozi.jpg" />
+                    <img src="./img/ozi.jpg" />
                 </figure>
                 <div class="twitter__block-text">
                     <div class="name">おじいさん<span class="name_reply">@got</span></div>
