+++ src/js/models/userlogin.js
@@ -1,0 +1,10 @@
+// ユーザー登録
+export const postBtn = document.getElementById('post-btn');
+export const user_name=document.getElementById('UserName');
+export const user_bio=document.getElementById('User');
+export const user_mail=document.getElementById('Email');
+export const user_pass=document.getElementById('Password');
+export const user_confirmpass=document.getElementById('ConfirmPassword');
+
+
+
